+++ CHANGELOG.md
@@ -14,6 +14,7 @@ keyspace just as efficient as with ``SCAN``. (Note: The use of ``KEYS`` is still
 generally discouraged due to potentially huge response size, don't use in production)
 - Add ``RECOVERY-SCAN`` command for scanning through complete keyspace, including
 internal rocksdb keys.
+- Add tool ``quarkdb-sst-inspect`` to allow low-level inspection of SST files.
 
 ## 0.3.8 (2019-05-27)
 - Prevent elections from hanging on the TCP timeout when one of the member hosts